remove whitelsit check, make auth gate only
