+++ app/src/renderer/src/main.tsx
@@ -12,7 +12,7 @@ import { TTSProvider } from './lib/ttsProvider'
 import { GoServerProvider } from './contexts/GoServerContext'
 import { AuthProvider } from './contexts/AuthContext'
 import { routeTree } from '@renderer/routeTree.gen'
-import InvitationGate from './components/onboarding/InvitationGate'
+import AuthGate from './components/onboarding/AuthGate'
 import UpdateNotification from './components/UpdateNotification'
 import DependenciesGate from './components/setup/DependenciesGate'
 
@@ -39,9 +39,9 @@ function App() {
             <div className="flex flex-col h-screen w-screen bg-background">
               <DependenciesGate>
                 <AuthProvider>
-                  <InvitationGate>
+                  <AuthGate>
                     <RouterProvider router={router} />
-                  </InvitationGate>
+                  </AuthGate>
                 </AuthProvider>
               </DependenciesGate>
             </div>
